🚚 Move `scheduler` arbitrary into /arbitrary
