+++ src/arbitrary/_internals/SchedulerArbitrary.ts
@@ -1,0 +1,35 @@
+import { NextArbitrary } from '../../check/arbitrary/definition/NextArbitrary';
+import { NextValue } from '../../check/arbitrary/definition/NextValue';
+import { Random } from '../../random/generator/Random';
+import { Stream } from '../../stream/Stream';
+import { Scheduler } from './interfaces/Scheduler';
+import { ScheduledTask, SchedulerImplem } from './implementations/SchedulerImplem';
+
+/** @internal */
+export class SchedulerArbitrary<TMetaData> extends NextArbitrary<Scheduler<TMetaData>> {
+  constructor(readonly act: (f: () => Promise<void>) => Promise<unknown>) {
+    super();
+  }
+
+  generate(mrng: Random, _biasFactor: number | undefined): NextValue<Scheduler<TMetaData>> {
+    const buildNextTaskIndex = (r: Random) => {
+      return {
+        clone: () => buildNextTaskIndex(r.clone()),
+        nextTaskIndex: (scheduledTasks: ScheduledTask<TMetaData>[]) => {
+          return r.nextInt(0, scheduledTasks.length - 1);
+        },
+      };
+    };
+    return new NextValue(new SchedulerImplem<TMetaData>(this.act, buildNextTaskIndex(mrng.clone())));
+  }
+
+  canGenerate(value: unknown): value is Scheduler<TMetaData> {
+    // Not supported yet
+    return false;
+  }
+
+  shrink(_value: Scheduler<TMetaData>, _context?: unknown): Stream<NextValue<Scheduler<TMetaData>>> {
+    // Not supported yet
+    return Stream.nil();
+  }
+}
